Remove refactor_me.Models namespace
