+++ refactor-me/Controllers/ProductsController.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Net;
 using System.Web.Http;
-using refactor_me.Models;
 using refactor_me.Model.Contract;
 
 namespace refactor_me.Controllers
