+++ refactor-me/Repository/BaseRepository.cs
@@ -1,5 +1,4 @@
-﻿using refactor_me.Models;
-using System;
+﻿using System;
 using System.Data;
 using System.Data.SqlClient;
