+++ refactor-me/Repository/Helpers.cs
@@ -1,7 +1,7 @@
 ﻿using System.Data.SqlClient;
 using System.Web;
 
-namespace refactor_me.Models
+namespace refactor_me.Repository
 {
     public class Helpers
     {
